separate models out and add tests
